Add MPL license, coding style and commit message guidelines.
The license is MPL 2.0 as required by GENIVI.

The coding style is adopted from GLib-based projects such as Clutter
but has been extended with best practices from the Xfce community.

The guidelines for git commit messages are taken from the GNOME
community among others and are commonly used in the open source world.