Drop intltool, depend on systemd 37 for now
Since the boot manager will run in the background as a service, we'll
probably not need translations. Systemd 183 is not available on all
Linux distributions yet, so in order to get started, we'll build on top
of 37 for now. 37 doesn't have all the features we need but it's good
enough to develop the LUC handler at least.