Fix args in XML interface spec, add example code to handle Register()
The XML element name for D-Bus method parameters is "arg", not "param".

luc_handler_service_handle_register() now has a "GVariant *apps"
parameter and reads/prints the foreground apps that are passed to it.
This is just example code to demonstrate how this can be done. It
will later be replaced with the real implementation.