Implement org.genivi.LUCHandler1.Register()
This commit implements the "handle-register" signal handler of the
LUCHandler interface skeleton by merging the current LUC and the
newly registered LUC types/apps and updating the
org.genivi.LUCHandler1:LastUserContext property.

It makes sure that the order in which LUC types are added to the final
GVariant as dictionary keys is always the same (alphabetical order).
This makes testing easier as D-Bus and GVariant both preserve the order
of dictionary keys (unlike with a GHashTable).

This commit also introduces common/glib-extensions.{c,h} where we can
dump functions that extend GLib in simple ways. The first helper
function is g_variant_string_array_has_string() which we use to check if
a string is included in a string array GVariant.