Implement org.genivi.LUCHandler1.Deregister()
This commit implements the "handle-deregister" signal handler of the
LUCHandler interface skeleton. It essentially copies the current last
user context and drops all the apps that are part of the Deregister()
call. It then updates the org.genivi.LUCHandler1:LastUserContext
property.

The deregistration algorithm makes sure that LUC types for which no apps
are registered anymore are dropped from the LUC entirely.