Add --register command-line option to LegacyAppHandler
When legacy-app-handler is called without arguments, it starts as a
service, and as the primary instance of the LegacyAppHandler
GApplication.

When legacy-app-handler is called with arguments, it is started as a
remote instance of the LegacyAppHandler GApplication, which passes its
command-line arguments to the primary instance.

On receiving command-line arguments from a remote instance, the Primary
instance will parse those arguments, expecting a command like the
following:
  legacy-app-handler --register --unit="foo" --timeout=120 \
  --shutdown-mode="normal"

Timeout and shutdown-mode are optional, and if omitted will default to
0, and "normal", respectively.

Because interaction with the Node State Manager is not yet implemented,
the Legacy App Handler will print the arguments to DLT, instead of
registering the app as a shutdown consumer.