Define D-Bus interface for Shutdown Consumers
Shutdown Consumers are D-Bus objects that exist for every application
that does not have a Shutdown() method, and are registered with the Node
State Manager (NSM).

When the NSM shuts down, it calls Shutdown() to all Shutdown Consumers
in reverse order to their registration. For legacy applications, a
Shutdown Consumer is created and registered with the NSM.

When the Shutdown Consumer is told to shut down, it shuts down the
legacy application by signalling the Boot Manager.