Change legacy app handler to register on the system bus
Makes the legacy app handler connect to the system bus instead of the
session bus.

This commit also changes bus names to consistently use LegacyAppHandler1
instead of LAHandler1.

This commit does not allow communication with the legacy app handler by
command-line. The following commit will provide this.

This commit does not provide a way to start the legacy app handler
through systemd. The following commit also provides this.