Better test case titles and always set the node state in TSM tests
Test titles are made more descriptive by including the name of the node
state which is being set.

A link is included to the enum which defines the NSMNodeState.

Tests are changed to always set the node state. This is done by
starting the unit twice in most cases. In addition, the less-important
lines in the DLT log output are no longer specified.

Capitalization in titles has been cleaned up somewhat.