Rewrite documentation for legacy app handler test
Removed the requirement to restart the system for each test, instead
just restart the node startup controller for each test. The downside of
this is that the dlt.log file is not reset so may become very large.

Reduced the number of DLT log messages to compare against the output.
Instead, only the most relevant lines are included for each test.

Added separate test that the Node Startup Controller registers itself.

Rewrote the descriptions for each test case to be more readable.

Numbered each test case.

Spelling and grammar fixes.