Remove the list of units and create D-Bus proxy only for wanted targets
Remove the list of units which is not longer used.
When the TargetStarterMonitor receives a "job-removed" signal it
creates a D-Bus proxy to monitor the unit and check if the state
changes to "active". If it does the TargetStartupMonitor sets the node
state.

The TargetStarterMonitor is only interested in monitoring
"focussed.target", "unfocussed.target" and "lazy.target" units. To
avoid creating the D-Bus proxy for unneeded units, it should check in
target_startup_monitor_job_removed if the unit is a unit which has to be
monitored.